refactor(evals): extract evaluator service from eval-suites.service.ts
eval-suites.service.ts grew to 1721 lines. Extract evaluator-specific
logic into eval-suites-evaluators.service.ts (~600 lines):
- resolveAssertions (merge logic)
- createEvaluator, deleteSuiteEvaluator (suite-level CRUD)
- createTestCaseEvaluator, getTestCaseEffectiveEvaluators,
  deleteTestCaseEvaluator (override CRUD)
- loadSuiteEvaluators (shared helper)

Re-exports from original file so no callers need updating.
eval-suites.service.ts is now ~1120 lines.
